main()을 다른 클래스로 분리
기능을 분리했다.
