+++ src/mission/controller/LadderGameController.java
@@ -1,0 +1,28 @@
+package mission.controller;
+
+import mission.model.Ladder;
+import mission.view.InputView;
+import mission.view.OutputView;
+
+import java.io.IOException;
+
+public class LadderGameController {
+        InputView inputView;
+
+        Ladder ladder;
+
+        OutputView outputView;
+
+        public LadderGameController() {
+                this.inputView = new InputView();
+                this.ladder = new Ladder();
+                this.outputView = new OutputView();
+        }
+
+        public void run() throws IOException {
+                String[] peopleNames = inputView.getPeopleName();
+                int ladderHeight = inputView.getLadderHeight();
+                ladder.makeLadder(peopleNames.length, ladderHeight);
+                outputView.printPeopleNameAndLadder(peopleNames,ladderHeight, ladder.getLadder());
+        }
+}
